Board::writeMotor() takes floating-point value

+++ src/board.hpp
@@ -31,7 +31,7 @@ class Board {
         virtual void     delayMilliseconds(uint32_t msec) = 0;
         virtual void     getImu(float eulerAnglesRadians[3], int16_t gyroRaw[3]) = 0;
         virtual uint64_t getMicros() = 0;
-        virtual void     writeMotor(uint8_t index, uint16_t value) = 0;
+        virtual void     writeMotor(uint8_t index, float value) = 0;
 
     //------------------------------------------ LED ------------------------------------------------------------
         virtual void     ledSet(uint8_t id, bool is_on) { (void)id; (void)is_on; }
